[Task-43] Updated CsvtoJson, CsvHandler, and Main Functions
