--- csv_handler/csv/csv.go
@@ -1,25 +1,0 @@
-// write parseCsv function
-
-package csv
-
-import (
-	"encoding/csv"
-	"os"
-)
-
-func ParseCsv(filePath string) ([][]string, error) {
-	file, err := os.Open(filePath)
-	if err != nil {
-		return nil, err
-	}
-	defer file.Close()
-
-	reader := csv.NewReader(file)
-
-	records, err := reader.ReadAll()
-	if err != nil {
-		return nil, err
-	}
-
-	return records, nil
-}
