--- mycsv/record.csv
@@ -1,4 +1,0 @@
-ID, Name, Occupation
-1, Zeeshan, Software Engineer
-2, Nauman, Lead Software Engineer
-3, Haris, Front End Developer
